+++ nexus/examples/dvr/recpump_with_wrap.c
@@ -1,7 +1,7 @@
 /******************************************************************************
- *    (c)2008-2011 Broadcom Corporation
+ * Copyright (C) 2008-2018 Broadcom. The term "Broadcom" refers to Broadcom Limited and/or its subsidiaries.
  *
- * This program is the proprietary software of Broadcom Corporation and/or its licensors,
+ * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
  * conditions of a separate, written license agreement executed between you and Broadcom
  * (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
@@ -34,17 +34,6 @@
  * ACTUALLY PAID FOR THE SOFTWARE ITSELF OR U.S. $1, WHICHEVER IS GREATER. THESE
  * LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF ESSENTIAL PURPOSE OF
  * ANY LIMITED REMEDY.
- *
- * $brcm_Workfile: $
- * $brcm_Revision: $
- * $brcm_Date: $
- *
- * Module Description:
- *
- * Revision History:
- *
- * $brcm_Log: $
- * 
  *****************************************************************************/
 #include "nexus_platform.h"
 #include "nexus_pid_channel.h"
@@ -73,8 +62,8 @@ static void print_status(NEXUS_RecpumpHandle recpump)
     NEXUS_RecpumpStatus status;
     NEXUS_Recpump_GetStatus(recpump, &status);
     printf("status: \n");
-    printf("  data:  %u total bytes \tfifo %u/%u\n", (unsigned)status.data.bytesRecorded, status.data.fifoDepth, status.data.fifoSize);
-    printf("  index: %u total bytes \tfifo %u/%u\n", (unsigned)status.index.bytesRecorded, status.index.fifoDepth, status.index.fifoSize);
+    printf("  data:  %u total bytes \tfifo %u/%u\n", (unsigned)status.data.bytesRecorded, (unsigned)status.data.fifoDepth, (unsigned)status.data.fifoSize);
+    printf("  index: %u total bytes \tfifo %u/%u\n", (unsigned)status.index.bytesRecorded, (unsigned)status.index.fifoDepth, (unsigned)status.index.fifoSize);
 }
 
 int main(void) {
@@ -175,7 +164,7 @@ int main(void) {
             }
             NEXUS_Recpump_IndexReadComplete(recpump, n+n2);
         }
-        printf("wrote %d+%d data, %d+%d index\n", data_buffer_size[0], data_buffer_size[1], index_buffer_size[0], index_buffer_size[1]);
+        printf("wrote %u+%u data, %u+%u index\n", (unsigned)data_buffer_size[0], (unsigned)data_buffer_size[1], (unsigned)index_buffer_size[0], (unsigned)index_buffer_size[1]);
         print_status(recpump);
     }
     /* loops forever */
