+++ nexus/examples/transport/tpit.c
@@ -1,7 +1,7 @@
 /******************************************************************************
- *    (c)2008-2010 Broadcom Corporation
+ * Copyright (C) 2008-2018 Broadcom. The term "Broadcom" refers to Broadcom Limited and/or its subsidiaries.
  *
- * This program is the proprietary software of Broadcom Corporation and/or its licensors,
+ * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
  * conditions of a separate, written license agreement executed between you and Broadcom
  * (an "Authorized License").  Except as set forth in an Authorized License, Broadcom grants
@@ -34,17 +34,6 @@
  * ACTUALLY PAID FOR THE SOFTWARE ITSELF OR U.S. $1, WHICHEVER IS GREATER. THESE
  * LIMITATIONS SHALL APPLY NOTWITHSTANDING ANY FAILURE OF ESSENTIAL PURPOSE OF
  * ANY LIMITED REMEDY.
- *
- * $brcm_Workfile: $
- * $brcm_Revision: $
- * $brcm_Date: $
- *
- * Module Description:
- *
- * Revision History:
- *
- * $brcm_Log: $
- * 
  *****************************************************************************/
 #include "nexus_platform.h"
 #include "nexus_pid_channel.h"
@@ -132,7 +121,7 @@ int main(void) {
         rc = NEXUS_Recpump_GetIndexBuffer(recpump, &buffer, &indexsize);
         BDBG_ASSERT(!rc);
         if (indexsize) {
-            printf("%d tpit bytes\n", indexsize);
+            printf("%u tpit bytes\n", (unsigned)indexsize);
             rc = NEXUS_Recpump_IndexReadComplete(recpump, indexsize);
             BDBG_ASSERT(!rc);
         }
